rename paypal provider enum file
