+++ packages/react-paypal-js/src/v6/context/PayPalProviderContext.tsx
@@ -3,7 +3,7 @@ import { createContext } from "react";
 import {
     INSTANCE_LOADING_STATE,
     INSTANCE_DISPATCH_ACTION,
-} from "../types/PayPalProviderEnums";
+} from "../types/ProviderEnums";
 
 import type {
     Components,
