+++ packages/react-paypal-js/src/v6/types/index.ts
@@ -1,4 +1,4 @@
-export * from "./PayPalProviderEnums";
+export * from "./ProviderEnums";
 export * from "./braintree";
 
 export type * from "@paypal/paypal-js/sdk-v6";
